split plugin test
